+++ CHANGES.rst
@@ -92,6 +92,7 @@ Version 3.2.0
 -   ``generate_password_hash`` uses ``secrets.token_urlsafe`` to generate salt.
     The private ``gen_salt`` method is removed. :pr:`3167`
 -   The test ``Client`` has a ``query`` method for the ``QUERY`` request method.
+-   ``unquote_etag`` and ``ETags.from_header`` discard invalid unquoted values.
 
 
 Version 3.1.9